~ pass args using decorators

+++ dot.py
@@ -6,6 +6,37 @@ from __future__ import unicode_literals
 import os, sys
 from os.path import join, isfile, isdir
 
+
+def option(command, default='__special_none__', mode=str):
+    """Decorator to pass sys.argv to functions,  
+    imitates https://github.com/pallets/click
+    """
+    if type(mode) != type:
+        raise ValueError('mode must be type')
+    opt = default
+    if command in sys.argv:
+        li = []
+        for i in sys.argv[sys.argv.index(command) + 1:]:
+            if i.find('-') == 0:
+                break
+            li.append(i)
+        if mode == tuple:
+            opt = tuple(li)
+        elif len(li) > 0:
+            if mode == bool:
+                opt = True
+            else:
+                opt = mode(li[0])
+        elif default == '__special_none__':
+            print('Missing value for %s.' % command)
+            sys.exit(-1)
+    def decorate(fn):
+        def wrapper(*args):
+            return fn(*args + (opt,))
+        return wrapper
+    return decorate
+
+
 help_text = """DOT:
     Naive dotfiles management util by geneLocated
 
@@ -20,42 +51,37 @@ commands:
     apply <topic>:	Apply a topic of .files (by making soft links in some directory)
     recover <topic>:	Stop manage a topic of .files, and try to restore
 
-    add <file> ...:	Stage files you want to backup to a temporary buffer
+    add <file> ...:	Stage files you want to backup in buffer
     status:		Display files you staged and the topic you are going to commit to
-    commit <topic>: 	 Store buffer data to a topic, you may want to run `{fn} apply <topic>` to make links in your directories
+    commit <topic>: 	Store buffer data to a topic
 
 operation flow:
-    to get files into a fresh management:
+    to get files into management:
         [{fn} add] -> ... -> [{fn} add] -> [{fn} commit] -> [{fn} apply] -> [git add .] -> [git commit]
-    to continue managing on maybe some other device:
-        [{fn} apply]
-
-HINT:
-    These files and dirctories are temporary. Instead of tracking them in your repo, add them in `.gitignore`:
-    /BUFFER/""".format(fn=sys.argv[0])
+    to continue managing:
+        [{fn} apply]""".format(fn=sys.argv[0])
 
 
 def help_():
     print(help_text)
 
 
-def list_():
-    if len(sys.argv) < 3:
+@option('list', default=None)
+def list_(topic=None):
+    if topic == None:
         # list topics in this repo
         print(_gettopic())
     else:
         # list dotfiles under a topic
-        print(_getfiles(_prefixtopic(sys.argv[2])))
+        print(_getfiles(_prefixtopic(topic)))
 
 
+# os.path.join('/opt/someplace', '/home/user/') returns the latter
 def _rmrootslash(dir_):
-    """It is impossible to join '/opt/someplace' and '/home/user/' together
-    using os.path.join. We must remove one the root splash of two dirs.
-    """
+    """Remove root slash"""
     if dir_[0] == '/':
         return dir_[1:]
-    else:
-        raise ValueError('do not contain root slash')
+    return dir_
 
 
 def _prefixtopic(topic):
@@ -83,32 +109,36 @@ def _getfiles(topic):
     return files
 
 
-def apply():
+@option('apply')
+def apply(topic):
     """Apply a topic."""
-    if len(sys.argv) < 3:
-        print('Missing topic, operation failed.')
-    else:
-        topic = _prefixtopic(sys.argv[2])
-        for target in _getfiles(topic):
-            ori = join(os.getcwd(), topic, _rmrootslash(target))
-            print(ori)
-            print(target)
-            # make dir if not exist
-            dir_ = _getdir(target)
-            if not isdir(dir_):
-                os.system('mkdir -p ' + dir_)
-            # backup the file if already exist
-            if isfile(target) and not isfile(target + '.BAK'):
-                os.rename(target, target + '.BAK')
-            elif isfile(target):
+    topic = _prefixtopic(topic)
+    for target in _getfiles(topic):
+        ori = join(os.getcwd(), topic, _rmrootslash(target))
+        print(ori)
+        print('\033[1;32m{}\033[0m'.format(target))
+        # make dir if not exist
+        dir_ = _getdir(target)
+        if not isdir(dir_):
+            os.system('mkdir -p ' + dir_)
+        # backup the file if already exist
+        if isfile(target) and not isfile(target + '.BAK'):
+            os.rename(target, target + '.BAK')
+        else:
+            try:
                 os.remove(target)
-            os.system('ln -s {ori} {tar}'.format(ori="'{}'".format(ori), tar=target))
+            # except FileNotFoundError:
+            except OSError:
+                pass
+        os.system('ln -s {ori} {tar}'.format(ori="'{}'".format(ori), tar=target))
 
 
-def recover():
+@option('recover')
+def recover(topic):
     """Remove the link file, and recover from the .BAK file"""
-    topic = _prefixtopic(sys.argv[2])
+    topic = _prefixtopic(topic)
     for f in _getfiles(topic):
+        f = _rpenv(f)
         if isfile(f):
             os.remove(f)
         if isfile(f + '.BAK'):
@@ -125,47 +155,42 @@ def _getshortname(fullname):
     return os.path.split(fullname)[1]
 
 
-def env_add(string):
-    from shutil import copy
+def _rpenv(string):
+    """Replace env var with its value."""
     import re
-    found = re.findall('{.*?}', string)
-    fn_value = string  # /home/user/.vimrc like
-    fn_var = string  # ${HOME}/.vimrc like
+    found = re.findall('\\${.*?}', string)
+    if len(found) == 0:
+        return string
+    fn_value = string  # /home/user/.vimrc
     # bracketed environment variable
     for bkt_var in found:
-        fn_var = fn_var.replace(bkt_var, '$' + bkt_var)
-        var = bkt_var[1:-1]  # remove {} brackets
+        var = bkt_var[2:-1]  # remove ${}
         if var in os.environ:
             fn_value = fn_value.replace(bkt_var, os.environ[var])
         else:
             raise ValueError(
                 'environment variable `${}` not exist'.format(bkt_var))
-    after = _getdir(join('BUFFER', fn_var))
-    if not isdir(after):
-        os.makedirs(after)
-    copy(fn_value, after)
+        return fn_value
 
 
-def add():
-    """Add file(s) into the buffer."""
+@option('add', mode=tuple)
+def add(files):
+    """Add file(s) into buffer."""
     from shutil import copy
-    # environment variable mode
-    if ('-e' in sys.argv) or ('--env' in sys.argv):
-        argv = sys.argv
-        argv.remove('-e') if '-e' in argv \
-            else argv.remove('--env')
-        for arg in argv[2:]:
-            env_add(arg)
-        return
-
-    if len(sys.argv) == 2:
+    if len(files) == 0:
         print('Nothing specified, nothing added.')
     else:
-        for arg in sys.argv[2:]:
-            after = _getdir(join('BUFFER', _rmrootslash(arg)))
+        for arg in files:
+            before = _rpenv(arg)
+            before = os.path.abspath(before)
+            after = arg
+            # add relative file to abs path BUFFER
+            if not os.path.isabs(_rpenv(arg)):
+                after = os.path.abspath(arg)
+            after = _getdir(join('BUFFER', _rmrootslash(after)))
             if not isdir(after):
                 os.makedirs(after)
-            copy(arg, after)
+            copy(before, after)
 
 
 def status():
@@ -176,16 +201,15 @@ def status():
         print('Nothing in the buffer.')
 
 
-def commit():
-    """Commit the files in the buffer to the selected topic.
-    Create one if not exist.
-    """
+@option('commit')
+def commit(topic):
+    """Store the files in the buffer to a topic."""
     if not (isdir('BUFFER') and len(_getfiles('BUFFER')) != 0):
         print('Fatal: no file staged in the buffer')
         print('  (use `{} add <file> ...` to stage files)'.format(sys.argv[0]))
     else:
         from shutil import move
-        move('BUFFER', _prefixtopic(sys.argv[2]))
+        move('BUFFER', _prefixtopic(topic))
 
 
 if __name__ == '__main__':
@@ -195,14 +219,9 @@ if __name__ == '__main__':
         help_()
     else:
         {'help': help_,
-
          'list': list_,
-         'ls': list_,
          'apply': apply,
          'recover': recover,
-
          'add': add,
          'status': status,
-         'st': status,
-         'commit': commit,
-         'ci': commit}[sys.argv[1]]()
+         'commit': commit}[sys.argv[1]]()
